docs(cmd): change sub module directory names
these submodules act as commands to be executed by the git-xargs tool. Although rarely used, the directory names should at least make sense,
