+++ cmd/bump-github-actions/go.mod
@@ -1,0 +1,5 @@
+module update-tf
+
+go 1.19
+
+require gopkg.in/yaml.v3 v3.0.1
