--- cmd/bumpGHA/.github/workflows/unit.yml
@@ -1,24 +1,0 @@
-"on":
-    pull_request:
-        types:
-            - opened
-            - edited
-            - reopened
-            - synchronize
-jobs:
-    go-tests:
-        name: Run Terratest Unit Tests
-        runs-on: ubuntu-latest
-        steps:
-            - name: Checkout
-              uses: actions/checkout@v2
-            - name: Setup Terraform
-              uses: hashicorp/setup-terraform@v1
-              with:
-                terraform_version: 1.2.5
-                terraform_wrapper: false
-            - name: Run tf Tests
-              working-directory: test/unit-test
-              run: |
-                terraform init
-                terraform validate
