--- cmd/bumpGHA/go.mod
@@ -1,5 +1,0 @@
-module update-tf
-
-go 1.19
-
-require gopkg.in/yaml.v3 v3.0.1
