+++ cmd/get-github-links/repos.txt
@@ -1,0 +1,50 @@
+cloud-platform-terraform-ecr-credentials
+cloud-platform-terraform-s3-bucket
+cloud-platform-terraform-rds-instance
+cloud-platform-terraform-elasticache-cluster
+cloud-platform-terraform-dynamodb-cluster
+cloud-platform-terraform-sqs
+cloud-platform-terraform-dms
+cloud-platform-terraform-elasticsearch
+cloud-platform-terraform-bastion
+cloud-platform-terraform-auth0
+cloud-platform-terraform-monitoring
+cloud-platform-terraform-logging
+cloud-platform-terraform-starter-pack
+cloud-platform-terraform-opa
+cloud-platform-terraform-ingress-controller
+cloud-platform-terraform-certmanager
+cloud-platform-terraform-template
+cloud-platform-terraform-external-dns
+cloud-platform-terraform-velero
+cloud-platform-terraform-cluster-autoscaler
+cloud-platform-terraform-concourse
+cloud-platform-terraform-awsaccounts-iam
+cloud-platform-terraform-awsaccounts-baselines
+cloud-platform-terraform-kops
+cloud-platform-terraform-kuberos
+cloud-platform-terraform-sonarqube
+cloud-platform-terraform-pushgateway
+cloud-platform-terraform-rds-aurora
+cloud-platform-terraform-flow-logs
+cloud-platform-terraform-serviceaccount
+cloud-platform-terraform-sns-topic
+cloud-platform-terraform-gatekeeper
+cloud-platform-terraform-irsa
+cloud-platform-terraform-eks-csi
+cloud-platform-terraform-amp
+cloud-platform-terraform-eks-add-ons
+cloud-platform-terraform-github-prototype
+cloud-platform-terraform-alb-ingress-controller
+cloud-platform-terraform-aws-sso
+cloud-platform-terraform-global-resources-auth0
+cloud-platform-terraform-kuberhealthy
+cloud-platform-terraform-secrets-manager
+cloud-platform-terraform-efs-csi
+cloud-platform-terraform-efs-pv
+cloud-platform-terraform-aws-scheduler
+cloud-platform-terraform-secrets-store-csi
+cloud-platform-terraform-descheduler
+cloud-platform-terraform-canary
+cloud-platform-terraform-cloudfront
+cloud-platform-terraform-trivy-operator
