+++ cmd/upgrade-teraform-version/go.mod
@@ -1,0 +1,20 @@
+module upgradeTerraform
+
+go 1.19
+
+require (
+	github.com/fatih/color v1.14.1
+	github.com/hashicorp/hcl/v2 v2.16.1
+	github.com/zclconf/go-cty v1.12.1
+)
+
+require (
+	github.com/agext/levenshtein v1.2.1 // indirect
+	github.com/apparentlymart/go-textseg/v13 v13.0.0 // indirect
+	github.com/google/go-cmp v0.3.1 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.17 // indirect
+	github.com/mitchellh/go-wordwrap v0.0.0-20150314170334-ad45545899c7 // indirect
+	golang.org/x/sys v0.3.0 // indirect
+	golang.org/x/text v0.3.7 // indirect
+)
